brush: refactor to keep noncomponents files on camelCase
