+++ scissors/src/fromSchemeGetTileNameById.js
