+++ scissors/src/visionManager.js
